Rename: src/generated/{factory => factories}/*
`factories` is the name of the folder created by the hardhat compiler.
Using the same name allows easier code sharing betwee the open source
and the internal repositories.
